Add alias for bypassing global pip requiring venv
I want to have the safety rail of prohibiting global installs for normal
pip, but this makes it annoying when I do want to install globally. The
gpip alias makes it so that I can explicitly specify a "global" version
of the pip command without having to remember the specific environment
variable syntax.